backend/requirements.txt
fastapi
uvicorn
streamlit
requests
openai
python-dotenv
Pillow
matplotlib